Change rowEchelon() and rowReducedEchelon()
Switch row-reduced to row-echelon

fixed "row-reduced" to "row-echelon" in some comments before back-substitution, particularly in xVec() and rowEchelon()
changed `dimension` to `temp.dimension` in rowEchelon() method
removed code in rowEchelon() that didn't appear to do anything; specifically the double for-loop at the end of the method. 

It was supposed to "place the pivots where they belong" but it didn't appear to work. Removed the same code from xVec().

Make getDeterminant() more efficient: if the multiplier is zero, then minor is zero.

Issues:
1) On some matrices, i.e. those with a zero first row and zero first column, the matrix leaves the zero row at the top.
2) On matrices with some rows having elements equal to zero up to some element, error.

Solution: Put all zero rows at the bottom prior to beginning the procedure. Treat the matrix with first columns that are 

zero as if it the column wasn't there; that is, form a (n)x(n-1) of the original matrix without the first column, if it's 

a zero column.

3. On rowReducedEchelon, a problem occurs in the for loop: Implies the leading number in each row
 is a pivot; this may not 

be the case, however.

How to fix? Find the pivot columns. For every pivot column, find all numbers to the top of the pivot (in the same column) that are non-zero. Subtract those rows containing non-zero entries.

Add changeToZerosIfNeeded() to rowReducedEchelon().

Fix the inconsistency of having `dimension` in loop when it should be `temp.dimension` inside rowEchelon().

Added a "zero-rows" check to rowReducedEchelon.

Added more comments to xVec()'s if, else if statements at the top.

Some inconsistencies, such as using `temp.matrix[i][j]` vs `temp(i + 1, j + 1)` remain in the code in other methods. I 

would prefer the former in the code.

A problem was also found when checking if the augmented matrix is inconsistent in xVec() method. x_2 may be free but we 

will never know since it starts by checking if x_4 is free; if it's not, then it checks if x_4's equation will be 

inconsistent. It is impossible for the system to be inconsistent if there's a free variable.

Solution: Fix the definitions of inconsistency and free-variables.

The free-variable solution specifically requires you to start at the bottom, find the first non-zero row, find the first x 

at its index, then the next x at its index-1. if that x is free, go to index-2. repeat.

How did I solve it?

I assumed it was linear, such that we had to find x_5, then x_4, then x_3, then x_2, then x_1. Suppose x_2 and x_5 were 

found to be free. Then I would go to the first non-zero row, find x_5, but x_5 was found to be free, so my loop 

`continue`d to find x_4. Then we found x_3, then x_2, which it `continue`d, so we found x_1, and that was it.

Renamed `helperFunc()` to the more specific `isValidInput()`.

Added an override to `changeToZerosIfNeeded` for `vector` types.

Switched "back-substitution" to "creating zeros above each pivot" inside `rowReducedEchelon()`.